+++ src/models/submodels/WindowsOsquery.js
@@ -1,5 +1,6 @@
 import { run , runPower} from '../Osquery'
 import * as WindowsWMI from './WindowsWMI'
+import { isArray, isEmpty, isObject } from 'lodash'
 
 export const findRegistry = async ({ select = '*',  key , params = {}}) => {
     const { callback = (x) => x , error = (err) => {throw err}}  = params
@@ -84,4 +85,51 @@ export const getApplications = async ( params = {})  => {
     } catch (error) {
         error( e )
     }
+}
+
+export const getCallbackDevices = ( defValues ) => {
+    const callback = async ( result ) => {
+        let values =  await result
+        let defValuesFiltered = []
+        let valuesFiltered = []
+        values.forEach( ( details, key) => {
+            let filterValue = Changes.getIntersect( details
+                , ['Caption', 'CreationClassName', 'Description', 'DeviceID', 'PNPClass', 'Present', 'Service', 'Status'] 
+            );
+            valuesFiltered[ filterValue.DeviceID ] = filterValue ;
+        } )
+
+        defValues.forEach( ( details, key) => {
+            let filterValue = Changes.getIntersect( details
+                , ['Caption', 'CreationClassName', 'Description', 'DeviceID', 'PNPClass', 'Present', 'Service', 'Status'] 
+            );
+            defValuesFiltered[ filterValue.DeviceID] = filterValue ;
+        } )
+
+        let newValues = [], missingValues = []
+
+        Object.entries(defValuesFiltered).forEach( ( det, j ) => {
+            const [ i , ...detValues ] = det
+            const nValue = isObject( valuesFiltered[ i ] ) ? valuesFiltered[ i ] : []
+            isEmpty(nValue) ? missingValues.push( detValues ) : null
+        })
+
+        Object.entries(valuesFiltered).forEach( ( det, j ) => {
+            const [ i , ...detValues ] = det
+            const nValue = isObject( defValuesFiltered[ i ] ) ? defValuesFiltered[ i ] : []
+            isEmpty(nValue) ? newValues.push( detValues ) : null
+        })
+
+        if(newValues.length !== 0 || missingValues.length !== 0 ){
+            console.group()
+            console.assert( newValues?.length === 0 , newValues.length+' new values found...'.red)
+            console.assert( missingValues?.length === 0 , missingValues.length+' missing values found...'.red)
+            newValues?.length > 0 ? console.log( 'New Values', `${ Changes.inspect( newValues ) }`) : null
+            missingValues?.length > 0 ? console.log( 'Missing Values', `${ Changes.inspect( missingValues ) }`) : null
+            console.groupEnd()
+        } else {
+            console.log('No changes found...'.green )
+        }
+    }
+    return callback
 }
